NIFI-5566 Renamed new components to CryptographicHashAttribute and CryptographicHashContent. Added documentation about deprecated/renamed components. Fixed unit test resource path.
