+++ nifi-nar-bundles/nifi-standard-bundle/nifi-standard-processors/src/test/java/org/apache/nifi/processors/standard/TestHashContent.java
@@ -25,7 +25,7 @@ import org.apache.nifi.util.TestRunner;
 import org.apache.nifi.util.TestRunners;
 import org.junit.Test;
 
-public class TestHashContentLegacy {
+public class TestHashContent {
 
     @Test
     public void testMD5() throws IOException {
@@ -46,17 +46,17 @@ public class TestHashContentLegacy {
     }
 
     private void test(final String hashAlgorithm, final String expectedHash) throws IOException {
-        final TestRunner runner = TestRunners.newTestRunner(new HashContentLegacy());
-        runner.setProperty(HashContentLegacy.ATTRIBUTE_NAME, "hash");
-        runner.setProperty(HashContentLegacy.HASH_ALGORITHM, hashAlgorithm);
+        final TestRunner runner = TestRunners.newTestRunner(new HashContent());
+        runner.setProperty(HashContent.ATTRIBUTE_NAME, "hash");
+        runner.setProperty(HashContent.HASH_ALGORITHM, hashAlgorithm);
 
         runner.enqueue(Paths.get("src/test/resources/hello.txt"));
 
         runner.run();
         runner.assertQueueEmpty();
-        runner.assertAllFlowFilesTransferred(HashContentLegacy.REL_SUCCESS, 1);
+        runner.assertAllFlowFilesTransferred(HashContent.REL_SUCCESS, 1);
 
-        final MockFlowFile outFile = runner.getFlowFilesForRelationship(HashContentLegacy.REL_SUCCESS).get(0);
+        final MockFlowFile outFile = runner.getFlowFilesForRelationship(HashContent.REL_SUCCESS).get(0);
         final String hashValue = outFile.getAttribute("hash");
 
         assertEquals(expectedHash, hashValue);
